ended, responsive incomplete

+++ src/context/cart.tsx
@@ -1,3 +1,4 @@
+import { notifyToast } from "@/helpers/notify/notifyToast";
 import { ICartContext, IProduct } from "@/interfaces/products";
 import { createContext, ReactNode, useState } from "react";
 
@@ -26,13 +27,11 @@ export const CartProvider = ({ children }: { children: ReactNode }) => {
         localStorage.setItem('cart', JSON.stringify(newCart))
     }
     const handleDeleteQuantity = (product: IProduct) => {
-        console.log("borrando");
         if (product.quantity <= 1) {
             handleDeleteProduct(product)
             return
         }
         const newCart = cart.map(item => {
-            console.log("mapeando para borrar");
             
             if (item.name === product.name) {
                 return {
@@ -45,20 +44,21 @@ export const CartProvider = ({ children }: { children: ReactNode }) => {
         localStorage.setItem('cart', JSON.stringify(newCart))
     }
     const handleDeleteProduct = (product: IProduct) => {
-        const newCart = cart.filter(item => item.name !== product.name)
-        localStorage.setItem('cart', JSON.stringify(newCart))
-        handleGetCart()
+        notifyToast.confirmButton(()=> {
+            const newCart = cart.filter(item => item.name !== product.name)
+            localStorage.setItem('cart', JSON.stringify(newCart))
+            handleGetCart()
+        }, "¿Desea eliminar el producto del carrito?")
+        
     }
     const handleDeleteCart = () => {
         const newCart: IProduct[] = []
         localStorage.setItem('cart', JSON.stringify(newCart))
     }
     const handleAddToCart = async (product: IProduct) => {
-        console.log("agregando al carrito");
         
         const audio = new Audio('/dog-bark.mp3');
         audio.volume = 0.6
-        console.log("ladrido");
         
         await audio.play();
         if (cart.some(item => item.name === product.name)) {
@@ -70,7 +70,6 @@ export const CartProvider = ({ children }: { children: ReactNode }) => {
         handleGetCart()
     }
     const handleGetCart = () => {
-        console.log("obteniendo carrito");
         
         const cart = localStorage.getItem('cart') ? JSON.parse(localStorage.getItem('cart') as string) : []
         setCart(cart)
